requirements/requirements.txt
asttokens==3.0.1
attrs==26.1.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
easyocr==1.7.2
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.2
Flask==3.1.3
fsspec==2026.3.0
idna==3.11
ImageIO==2.37.3
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
lazy-loader==0.5
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
mpmath==1.3.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
networkx==3.6.1
ninja==1.13.0
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.2.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pyclipper==1.4.0
Pygments==2.20.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
setuptools==81.0.0
shapely==2.1.2
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
tifffile==2026.3.3
tinycss2==1.4.0
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
traitlets==5.14.3
triton==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
webencodings==0.5.1
Werkzeug==3.1.7
yarg==0.1.9
